+++ client/app/getDiabeticJournals/getGlucoseJournals/getGlucoseJournalsPage.tsx
@@ -12,6 +12,7 @@ import { useEffect, useState } from 'react';
 import { MdDeleteForever, MdInfoOutline, MdKeyboardArrowDown } from 'react-icons/md';
 import Header from '../../components/Header';
 import { formatDate, formatMilitaryTime } from '../../helpers/utils/datetimeformat';
+import Swal from 'sweetalert2';
 
 
 export default function GetGlucoseJournalsPage() {
@@ -47,11 +48,28 @@ export default function GetGlucoseJournalsPage() {
 
 
     async function deleteGlucoseJournals(glucoseJournalId: string){
-      const deleteresult = await deleteGlucoseJournal(glucoseJournalId);   
-      const newData = glucose && glucose.filter((item: { id: string; }) => item.id!=glucoseJournalId);
-      setglucose(newData);
-      router.push('/getDiabeticJournals');
-  
+        Swal.fire({
+          title: "Are you sure you want to delete this glucose journal entry?",
+          text: "You will not be able to retrieve it later",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Delete",
+          position: "bottom",
+        }).then(async (result: { isConfirmed: any; }) => {
+          if (result.isConfirmed) {
+            const deleteresult = await deleteGlucoseJournal(glucoseJournalId);   
+            const newData = glucose && glucose.filter((item: { id: string; }) => item.id!=glucoseJournalId);
+            setglucose(newData);
+            router.push('/getDiabeticJournals');
+            Swal.fire({
+              title: "Deleted!",
+              text: "Your glucose journal entry has been deleted.",
+              icon: "success"
+            });    
+          }
+        });
     }
