+++ client/app/getDiabeticJournals/getInsulinJournals/getInsulinJournalsPage.tsx
@@ -13,6 +13,7 @@ import { MdDeleteForever, MdInfoOutline, MdKeyboardArrowDown } from 'react-icons
 import Header from '../../components/Header';
 import Menu from '../../components/Menu';
 import { formatDate, formatMilitaryTime } from '../../helpers/utils/datetimeformat';
+import Swal from 'sweetalert2';
 
 export default function GetInsulinJournalsPage() {
   const logger = require('../../../logger');
@@ -47,10 +48,28 @@ export default function GetInsulinJournalsPage() {
 
 
     async function deleteInsulinJournals(insulinJournalId: string){
-      const deleteresult = await deleteInsulinJournal(insulinJournalId);   
-      const newData = insulin && insulin.filter((item: { id: string; }) => item.id!=insulinJournalId);
-      setinsulin(newData);
-      router.push('/getDiabeticJournals');
+      Swal.fire({
+        title: "Are you sure you want to delete this insulin journal entry?",
+        text: "You will not be able to retrieve it later",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Delete",
+        position: "bottom",
+      }).then(async (result: { isConfirmed: any; }) => {
+        if (result.isConfirmed) {
+          const deleteresult = await deleteInsulinJournal(insulinJournalId);   
+          const newData = insulin && insulin.filter((item: { id: string; }) => item.id!=insulinJournalId);
+          setinsulin(newData);
+          router.push('/getDiabeticJournals');
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your insulin journal entry has been deleted.",
+            icon: "success"
+          });    
+        }
+      });
     }
